test: Refactor, moving from src/ to test/ & splitting into three files
